rename eastings_to_graph to helpers
